+++ test_cards.py
@@ -1,4 +1,4 @@
-from cards import *
+from switch.cards import *
 
 VALUES = ['2', '3', '4', '5', '6', '7', '8', '9', '10', 'J', 'Q', 'K', 'A']
 SUITS = ['♣', '♢', '♡', '♠']